+++ doc/gb-model/release_notes.rst
@@ -13,6 +13,8 @@ Upcoming Release
 ================
 
 * Remove unnecessary output in `compose_networks` rule that causes error (#2)
+* Tabulated GSP wise powerplant capacities for GB (#4).
+* Tabulated EU country level aggregated powerplant capacities (#33)
 * Add rule 'retrieve_unavailability_data' to Snakemake workflow for fetching unavailability data from ENTSO-E. (#43)
 * Increase number of HTTP download retries to mitigate against Zenodo file retrieval timeouts.
 * Keep all retrieved data locally by default to reduce time spent re-downloading data on every run.